extract lineEventHandler() to a new file
