+++ handler/line_event_handler.js
@@ -1,0 +1,43 @@
+const { client } = require("../client/line_client");
+
+module.exports = function lineEventHandler(event) {
+    if (event.type !== 'message' || event.message.type !== 'text') {
+        // ignore non-text-message event
+        return Promise.resolve(null);
+    }
+
+    const lineID = event.source.userId;
+
+    const sentMessage = event.message.text;
+    var replyMessage = '';
+
+    if (sentMessage === 'book') {
+        const uri = `https://cal.com/nick-l-yang-vkljfs/15min?line-id=${lineID}`;
+        replyMessage = { type: 'text', text: `This is your booking URL: ${encodeURI(uri)}` };
+
+    } else {
+        // create an echoing text message
+        replyMessage = { type: 'text', text: `Hello "${lineID}" just said: "${sentMessage}"` };
+        console.log(`"${lineID}" just said: "${sentMessage}"`);
+
+        // send a notification after 5s
+        setTimeout(() => {
+            pushMessage(lineID);
+        }, 5000);
+
+    }
+
+    // use reply API
+    return client.replyMessage({
+        replyToken: event.replyToken,
+        messages: [replyMessage],
+    });
+}
+
+// push notification
+function pushMessage(userID) {
+    client.pushMessage({
+        to: userID,
+        messages: [{ type: 'text', text: 'Hello, This is a reminder message.' }],
+    });
+}
